break out drag and drop tests, add test
